+++ probpipe/__init__.py
@@ -88,6 +88,8 @@ from probpipe.core.event_template import (
     FunctionSpec,
     NumericEventTemplate,
     OpaqueSpec,
+    RecordSpec,
+    TermSpec,
     ValueSpec,
 )
 from probpipe.core.named_tree import NamedTree
@@ -289,6 +291,7 @@ __all__ = [
     "RecordBootstrapReplicateDistribution",
     "RecordDistribution",
     "RecordEmpiricalDistribution",
+    "RecordSpec",
     "SequentialJointDistribution",
     "SimpleGenerativeModel",
     "SimpleModel",
@@ -307,6 +310,7 @@ __all__ = [
     "SupportsUnnormalizedLogProb",
     "SupportsVariance",
     "TFPDistribution",
+    "TermSpec",
     "TrackedTerm",
     # Transformed
     "TransformedDistribution",